matching the template to the instruction in the content
