+++ docs/content/tutorial/connecting-to-external-services.mdx
@@ -37,7 +37,7 @@ You should use resources to manage communicating with external services because 
 
 Suppose your Hacker News pipeline in Dagster has picked up momentum, and your stakeholders want to learn more about who is using it. You are tasked with analyzing how many people sign up for Hacker News.
 
-In the scaffolded Dagster project you made in Part 2 of the tutorial, you may have noticed a directory called `resources` with some files in it. The most important file is the `resources.py` file that exposes a resource called `DataGeneratorResource`. This resource generates simulated data about Hacker News signups. You'll use this resource to get the data needed to produce an asset for analysis.
+In the scaffolded Dagster project you made in Part 2 of the tutorial, you may have noticed a directory called `resources` an `__init__.py` in it. It exposes a resource called `DataGeneratorResource`. This resource generates simulated data about Hacker News signups. You'll use this resource to get the data needed to produce an asset for analysis.
 
 <Note>
   The signup data from Hacker News is fake and generated by the library. This is
